rename ChatView to ChannelView
