+++ src/chat/components/ChannelView.tsx
